+++ src/main/java/tong/Ui.java
@@ -1,3 +1,7 @@
+package tong;
+
+import tong.command.CommandResult;
+
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
