+++ src/main/java/tong/command/Command.java
@@ -1,3 +1,7 @@
+package tong.command;
+
+import tong.TaskList;
+
 /**
  * Represents an executable command.
  */
