+++ src/main/java/tong/command/CommandResult.java
@@ -1,3 +1,5 @@
+package tong.command;
+
 import java.util.List;
 import java.util.Optional;
