+++ src/main/java/tong/command/ExitCommand.java
@@ -1,3 +1,5 @@
+package tong.command;
+
 /**
  * Terminates the program.
  */
