+++ src/main/java/tong/exception/DecodingException.java
@@ -1,3 +1,5 @@
+package tong.exception;
+
 public class DecodingException extends Exception {
     public DecodingException(String message) {
         super(message);
