+++ src/main/java/tong/exception/EmptyDescriptionException.java
@@ -1,3 +1,5 @@
+package tong.exception;
+
 public class EmptyDescriptionException extends Exception {
     public EmptyDescriptionException(String message) {
         super(message);
