+++ src/main/java/tong/exception/FolderNotFoundException.java
@@ -1,3 +1,5 @@
+package tong.exception;
+
 public class FolderNotFoundException extends Exception{
     public FolderNotFoundException(String message) {
         super(message);
