+++ src/main/java/tong/exception/TaskNotFoundException.java
@@ -1,0 +1,6 @@
+package tong.exception;
+public class TaskNotFoundException extends IndexOutOfBoundsException {
+    public TaskNotFoundException(String message) {
+        super(message);
+    }
+}
