+++ src/main/java/tong/task/Deadline.java
@@ -1,3 +1,5 @@
+package tong.task;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.LocalDateTime;
