+++ src/main/java/tong/task/Event.java
@@ -1,3 +1,5 @@
+package tong.task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
